Moved all test rules to their own package
