+++ app/src/androidTest/java/org/simple/clinic/help/HelpSyncAndroidTest.kt
@@ -8,7 +8,7 @@ import org.junit.Rule
 import org.junit.Test
 import org.junit.rules.RuleChain
 import org.junit.runner.RunWith
-import org.simple.clinic.LocalAuthenticationRule
+import org.simple.clinic.rules.LocalAuthenticationRule
 import org.simple.clinic.TestClinicApp
 import org.simple.clinic.util.RxErrorsRule
 import org.simple.clinic.util.unwrapJust
