+++ app/src/androidTest/java/org/simple/clinic/overdue/AppointmentSyncAndroidTest.kt
@@ -7,7 +7,7 @@ import org.junit.Before
 import org.junit.Rule
 import org.junit.rules.RuleChain
 import org.junit.runner.RunWith
-import org.simple.clinic.ServerAuthenticationRule
+import org.simple.clinic.rules.ServerAuthenticationRule
 import org.simple.clinic.TestClinicApp
 import org.simple.clinic.TestData
 import org.simple.clinic.patient.SyncStatus
