+++ app/src/androidTest/java/org/simple/clinic/reports/ReportsSyncAndroidTest.kt
@@ -8,7 +8,7 @@ import org.junit.Rule
 import org.junit.Test
 import org.junit.rules.RuleChain
 import org.junit.runner.RunWith
-import org.simple.clinic.ServerAuthenticationRule
+import org.simple.clinic.rules.ServerAuthenticationRule
 import org.simple.clinic.TestClinicApp
 import org.simple.clinic.util.None
 import org.simple.clinic.util.RxErrorsRule
